requirements.txt
asgiref==3.8.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
django==5.1.5
django-allauth==65.3.1
idna==3.10
oauthlib==3.2.2
pycparser==2.22
pyjwt==2.10.1
requests==2.32.3
requests-oauthlib==2.0.0
sqlparse==0.5.3
urllib3==2.3.0
django-extensions==3.2.3
pillow==11.1.0
django-browser-reload==1.17.0

